Make cmd+click to open a link in a new window work in Vivaldi
This broke when I switched to Vivaldi because I was depending on
undocumented Chrome behavior (the pendingUrl is present in Chrome
immediately upon tab creation). Now we monitor for tab updates and only
create the new Path OS window when the URL becomes available.

Since tabs may get additional updates after the Path OS window is
already created (I suspect this is because the update event fires when
the tab is removed, but I'm not sure) we have to keep track of the IDs
of tabs that have been moved to Path OS to avoid creating multiple Path
OS windows per new tab. I used a simple object-as-map to do this, but
this is a memory leak; the object grows without bound. A fixed-size
structure like a ring buffer would be more appropriate, but the leaky
approach works for now.